Move `Condvar` into `lock_api`
Adds new traits `RawCondvar` and `RawCondvarTimed` for implementing the inner state of a condvar, and moves `Condvar` to `lock_api` with a type parameter for the inner state. Also moves `WaitTimeoutResult`.

`parking_lot` then defines its own implementation as `RawCondvar`, which is publicly exported, along with `lock_api::{Condvar, WaitTimeoutResult}` with the type parameter filled. This mirrors how `Mutex` is exported.
