+++ src/lib.rs
@@ -17,6 +17,7 @@ mod elision;
 mod fair_mutex;
 mod mutex;
 mod once;
+mod raw_condvar;
 mod raw_fair_mutex;
 mod raw_mutex;
 mod raw_rwlock;
@@ -42,6 +43,7 @@ pub use self::condvar::{Condvar, WaitTimeoutResult};
 pub use self::fair_mutex::{const_fair_mutex, FairMutex, FairMutexGuard, MappedFairMutexGuard};
 pub use self::mutex::{const_mutex, MappedMutexGuard, Mutex, MutexGuard};
 pub use self::once::{Once, OnceState};
+pub use self::raw_condvar::RawCondvar;
 pub use self::raw_fair_mutex::RawFairMutex;
 pub use self::raw_mutex::RawMutex;
 pub use self::raw_rwlock::RawRwLock;
